[Review] Move objects from module
